+++ packages/cosmwasm-stargate/src/cosmwasmclient.ts
@@ -339,7 +339,7 @@ export class CosmWasmClient {
     return allCodes.map((entry: CodeInfoResponse): Code => {
       assert(entry.creator && entry.codeId && entry.dataHash, "entry incomplete");
       return {
-        id: entry.codeId.toNumber(),
+        id: Number(entry.codeId),
         creator: entry.creator,
         checksum: toHex(entry.dataHash),
       };
@@ -356,7 +356,7 @@ export class CosmWasmClient {
       "codeInfo missing or incomplete",
     );
     const codeDetails: CodeDetails = {
-      id: codeInfo.codeId.toNumber(),
+      id: Number(codeInfo.codeId),
       creator: codeInfo.creator,
       checksum: toHex(codeInfo.dataHash),
       data: data,
@@ -413,7 +413,7 @@ export class CosmWasmClient {
     assert(contractInfo.codeId && contractInfo.creator && contractInfo.label, "contractInfo incomplete");
     return {
       address: retrievedAddress,
-      codeId: contractInfo.codeId.toNumber(),
+      codeId: Number(contractInfo.codeId),
       creator: contractInfo.creator,
       admin: contractInfo.admin || undefined,
       label: contractInfo.label,
@@ -436,7 +436,7 @@ export class CosmWasmClient {
       assert(entry.operation && entry.codeId && entry.msg);
       return {
         operation: operations[entry.operation],
-        codeId: entry.codeId.toNumber(),
+        codeId: Number(entry.codeId),
         msg: JSON.parse(fromUtf8(entry.msg)),
       };
     });