+++ packages/cosmwasm-stargate/src/modules/wasm/aminomessages.ts
@@ -10,8 +10,7 @@ import {
   MsgStoreCode,
   MsgUpdateAdmin,
 } from "cosmjs-types/cosmwasm/wasm/v1/tx";
-import { AccessType } from "cosmjs-types/cosmwasm/wasm/v1/types";
-import Long from "long";
+import { AccessConfig, AccessType } from "cosmjs-types/cosmwasm/wasm/v1/types";
 
 export function accessTypeFromString(str: string): AccessType {
   switch (str) {
@@ -51,7 +50,7 @@ export function accessTypeToString(object: any): string {
 /**
  * @see https://github.com/CosmWasm/wasmd/blob/v0.18.0-rc1/proto/cosmwasm/wasm/v1/types.proto#L36-L41
  */
-export interface AccessConfig {
+export interface AminoAccessConfig {
   /**
    * Permission should be one kind of string 'Nobody', 'OnlyAddress', 'Everybody', 'AnyOfAddresses', 'Unspecified'
    * @see https://github.com/CosmWasm/wasmd/blob/v0.31.0/x/wasm/types/params.go#L54
@@ -77,7 +76,7 @@ export interface AminoMsgStoreCode {
     readonly sender: string;
     /** Base64 encoded Wasm */
     readonly wasm_byte_code: string;
-    readonly instantiate_permission?: AccessConfig;
+    readonly instantiate_permission?: AminoAccessConfig;
   };
 }
 
@@ -226,11 +225,11 @@ export function createWasmAminoConverters(): AminoConverters {
         sender: sender,
         wasmByteCode: fromBase64(wasm_byte_code),
         instantiatePermission: instantiate_permission
-          ? {
+          ? AccessConfig.fromPartial({
               permission: accessTypeFromString(instantiate_permission.permission),
               address: instantiate_permission.address ?? "",
               addresses: instantiate_permission.addresses ?? [],
-            }
+            })
           : undefined,
       }),
     },
@@ -260,7 +259,7 @@ export function createWasmAminoConverters(): AminoConverters {
         admin,
       }: AminoMsgInstantiateContract["value"]): MsgInstantiateContract => ({
         sender: sender,
-        codeId: Long.fromString(code_id),
+        codeId: BigInt(code_id),
         label: label,
         msg: toUtf8(JSON.stringify(msg)),
         funds: [...funds],
@@ -299,7 +298,7 @@ export function createWasmAminoConverters(): AminoConverters {
         fix_msg,
       }: AminoMsgInstantiateContract2["value"]): MsgInstantiateContract2 => ({
         sender: sender,
-        codeId: Long.fromString(code_id),
+        codeId: BigInt(code_id),
         label: label,
         msg: toUtf8(JSON.stringify(msg)),
         funds: [...funds],
@@ -368,7 +367,7 @@ export function createWasmAminoConverters(): AminoConverters {
       }: AminoMsgMigrateContract["value"]): MsgMigrateContract => ({
         sender: sender,
         contract: contract,
-        codeId: Long.fromString(code_id),
+        codeId: BigInt(code_id),
         msg: toUtf8(JSON.stringify(msg)),
       }),
     },