+++ packages/stargate/src/modules/gov/aminomessages.spec.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { TextProposal, VoteOption } from "cosmjs-types/cosmos/gov/v1beta1/gov";
 import { MsgDeposit, MsgSubmitProposal, MsgVote, MsgVoteWeighted } from "cosmjs-types/cosmos/gov/v1beta1/tx";
-import Long from "long";
 
 import { AminoTypes } from "../../aminotypes";
 import {
@@ -18,7 +17,7 @@ describe("AminoTypes", () => {
       const msg: MsgDeposit = {
         amount: [{ amount: "12300000", denom: "ustake" }],
         depositor: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
-        proposalId: Long.fromNumber(5),
+        proposalId: BigInt(5),
       };
       const aminoTypes = new AminoTypes(createGovAminoConverters());
       const aminoMsg = aminoTypes.toAmino({
@@ -73,7 +72,7 @@ describe("AminoTypes", () => {
     it("works for MsgVote", () => {
       const msg: MsgVote = {
         option: VoteOption.VOTE_OPTION_NO_WITH_VETO,
-        proposalId: Long.fromNumber(5),
+        proposalId: BigInt(5),
         voter: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
       };
       const aminoTypes = new AminoTypes(createGovAminoConverters());
@@ -94,7 +93,7 @@ describe("AminoTypes", () => {
 
     it("works for MsgVoteWeighted", () => {
       const msg: MsgVoteWeighted = {
-        proposalId: Long.fromNumber(5),
+        proposalId: BigInt(5),
         voter: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
         options: [
           { option: VoteOption.VOTE_OPTION_NO_WITH_VETO, weight: "700000000000000000" /* 0.7 */ },
@@ -135,7 +134,7 @@ describe("AminoTypes", () => {
       const expectedValue: MsgDeposit = {
         amount: [{ amount: "12300000", denom: "ustake" }],
         depositor: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
-        proposalId: Long.fromNumber(5),
+        proposalId: BigInt(5),
       };
       expect(msg).toEqual({
         typeUrl: "/cosmos.gov.v1beta1.MsgDeposit",
@@ -188,7 +187,7 @@ describe("AminoTypes", () => {
       const msg = new AminoTypes(createGovAminoConverters()).fromAmino(aminoMsg);
       const expectedValue: MsgVote = {
         option: VoteOption.VOTE_OPTION_NO_WITH_VETO,
-        proposalId: Long.fromNumber(5),
+        proposalId: BigInt(5),
         voter: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
       };
       expect(msg).toEqual({
@@ -211,7 +210,7 @@ describe("AminoTypes", () => {
       };
       const msg = new AminoTypes(createGovAminoConverters()).fromAmino(aminoMsg);
       const expectedValue: MsgVoteWeighted = {
-        proposalId: Long.fromNumber(5),
+        proposalId: BigInt(5),
         voter: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
         options: [
           { option: VoteOption.VOTE_OPTION_NO_WITH_VETO, weight: "750000000000000000" },